feat: allow to override how model metadata is loaded
