+++ Models/Models/ModelMetadataAttribute.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using System.IO;
+using System.Linq;
+using System.Reflection;
 
 namespace NMF.Models
 {
@@ -35,6 +38,41 @@ namespace NMF.Models
                 ModelUri = new Uri(modelUri, UriKind.Relative);
             }
             ResourceName = resourceName;
+        }
+
+        /// <summary>
+        /// Loads the manifest stream containing the metamodel from the given assembly and resource names. If the resource name is not found, an exception is thrown.
+        /// </summary>
+        /// <param name="assembly">The assembly to which this attribute is attached</param>
+        /// <param name="resourceNames">the resource names available for the given assembly</param>
+        /// <returns>A stream containing the metamodel described by this instance</returns>
+        public virtual Stream LoadMetamodel(Assembly assembly, string[] resourceNames)
+        {
+            var actualResourceName = FindResourceName(assembly, resourceNames);
+            return assembly.GetManifestResourceStream(actualResourceName);
+        }
+
+        private string FindResourceName(Assembly assembly, string[] names)
+        {
+            var resourceName = ResourceName;
+            if (!names.Contains(ResourceName))
+            {
+                var resources = names.Where(n => n.EndsWith(resourceName)).ToList();
+                if (resources.Count == 1)
+                {
+                    resourceName = resources[0];
+                }
+                else if (resources.Count == 0)
+                {
+                    throw new InvalidOperationException($"Embedded resource {resourceName} was not found in assembly {assembly.FullName}.");
+                }
+                else
+                {
+                    throw new InvalidOperationException($"Multiple embedded resources with the suffix {resourceName} were found in {assembly.FullName}.");
+                }
+            }
+
+            return resourceName;
         }
     }
 }
